requirements.txt
matplotlib==3.7.1
notion-client==2.0.0
pyyaml==6.0
